+++ src/Wyam.Core/Documents/Document.cs
@@ -294,6 +294,8 @@ namespace Wyam.Core.Documents
 
         IDocument IMetadata.Document(string key) => _metadata.Document(key);
 
+        public IReadOnlyList<IDocument> Documents(string key) => _metadata.Documents(key);
+
         public string Link(string key, string defaultValue = null, bool pretty = true) => _metadata.Link(key, defaultValue, pretty);
 
         public dynamic Dynamic(string key, object defaultValue = null) => _metadata.Dynamic(key, defaultValue);