+++ src/Wyam.Modules.CodeAnalysis.Tests/TestDocument.cs
@@ -84,6 +84,11 @@ namespace Wyam.Modules.CodeAnalysis.Tests
             return Get<IDocument>(key);
         }
 
+        public IReadOnlyList<IDocument> Documents(string key)
+        {
+            return Get<IReadOnlyList<IDocument>>(key);
+        }
+
         public string Link(string key, string defaultValue = null, bool pretty = true)
         {
             string value = Get<string>(key, defaultValue);